refactoring :: auth
