+++ src/controller/auth.ts
@@ -2,6 +2,7 @@ import express, { Request, Response } from 'express';
 import auth from '../service/auth';
 import { apiLimit } from '../middleware/limit';
 import { verifyJWT } from '../middleware/jwt';
+import { AuthenticatedRequest } from '../types';
 
 const app = express();
 
@@ -11,7 +12,7 @@ app.post('/signup', apiLimit, (req: Request, res: Response) => {
 app.post('/login', apiLimit, (req: Request, res: Response) => {
   auth.login(req, res);
 });
-app.post('/refresh', apiLimit, verifyJWT, (req: Request, res: Response) => {
+app.post('/refresh', apiLimit, verifyJWT, (req: AuthenticatedRequest, res: Response) => {
   auth.refresh(req, res);
 });
 app.post('/email', apiLimit, (req: Request, res: Response) => {
@@ -20,7 +21,7 @@ app.post('/email', apiLimit, (req: Request, res: Response) => {
 app.patch('/password', apiLimit, (req: Request, res: Response) => {
   auth.passwordModify(req, res);
 });
-app.post('/logout', apiLimit, verifyJWT, (req: Request, res: Response) => {
+app.post('/logout', apiLimit, verifyJWT, (req: AuthenticatedRequest, res: Response) => {
   auth.logout(req, res);
 });
